refactor: product module
