+++ src/renderer/utils/products/index.ts
@@ -1,3 +1,2 @@
 export * from './catalog';
-export * from './types';
 export * from './utils';
